requirements.txt
pystac-client>=0.7.7
planetary-computer>=1.0.0
rasterio>=1.3.9
rioxarray>=0.15.5
xarray>=2024.6.0
dask[complete]>=2024.5.0
geopandas>=0.14.4
shapely>=2.0.4
pyproj>=3.6.1
fiona>=1.9.6
pandas>=2.2.2
numpy>=1.26.4
scikit-image>=0.24.0
scikit-learn>=1.4.2
albumentations>=1.4.4
tqdm>=4.66.4
pyyaml>=6.0.1
rio-cogeo>=5.3.4
matplotlib>=3.8.4
# Install torch/torchvision separately for your platform:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121   # (CUDA 12.1)
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu    # CPU
